refactor: 딤처리해주는 함수를 util로 분리
검색바에서도 사이드바에서도 공통적으로 사용해야하는 함수라 util로 분리했다
