+++ src/js/main/event.js
@@ -1,0 +1,8 @@
+import { $, $All } from "../utils.js";
+
+export const addLeftBtnEventListeners = () => {
+    $(".banner-left-btn");
+};
+export const addRightBtnEventListeners = () => {
+    $(".banner-right-btn");
+};
